requirements-actual.txt
# 实际使用的依赖（自下而上开发）
# 更新于: 2025-01-07

opencv-python==4.8.1.78
compreface-sdk
python-dotenv
requests-toolbelt>=0.9.1
urllib3>=1.26
